Update README.md
My app is a simple yet effective way to buy and sell products online. It allows users to browse a wide variety of products, from clothes and electronics to books and home goods. Users can also create their own listings and sell their own products. The app is easy to use and navigate, and it offers a variety of features that make it a convenient and efficient way to shop online.

Here are some of the features of the app:

A wide variety of products to choose from
Easy to use and navigate interface
Ability to create and sell your own products
Secure payment processing
Fast and reliable shipping
The app is a great way to find the products you need and want, and it offers a convenient and efficient way to shop online.

Here is a more detailed description of the code:

The code is written in C++ and uses the following libraries:

iostream - for input and output
fstream - for file input and output
string - for strings
iomanip - for formatting output
The code defines a class called Item to represent an item in the store. The class has the following members:

itemCode - a string that uniquely identifies the item
itemName - a string that contains the name of the item
itemPrice - a double that contains the price of the item
The code also defines the following functions:

readItem() - reads the details of an item from the user
displayItem() - displays the details of an item
writeToFile() - writes the details of an item to a file
readFromFile() - reads the details of an item from a file
searchItemCode() - searches for an item by its code
The code is organized into the following functions:

main() - the main function of the program
readItem() - reads the details of an item from the user
displayItem() - displays the details of an item
writeToFile() - writes the details of an item to a file
readFromFile() - reads the details of an item from a file
searchItemCode() - searches for an item by its code